+++ stock_explorer/__init__.py
@@ -1,3 +1,3 @@
 """Modularer Aktien Explorer."""
 
-__version__ = "6.8.0"
+__version__ = "6.9.0"